notesappbackend/notesproject/requirements.txt
altgraph==0.17.2
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
auto-py-to-exe==2.15.0
beautifulsoup4==4.10.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==36.0.1
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
Eel==0.12.4
email-validator==1.2.1
filelock==3.7.1
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.7.1
MarkupSafe==2.1.1
outcome==1.1.0
pefile==2021.9.3
Pillow==9.1.0
platformdirs==2.5.2
pycparser==2.21
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.0
pyOpenSSL==22.0.0
pytz==2022.6
pywin32-ctypes==0.2.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.36
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.8
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
Werkzeug==2.1.2
whatsapp-web==0.0.1
whichcraft==0.6.1
wsproto==1.0.0
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
